code refactoring

+++ src/main/java/com/enviro/assessment/grad001/kganyamaleka/services/WasteCategoryService.java
@@ -2,6 +2,7 @@ package com.enviro.assessment.grad001.kganyamaleka.services;
 
 import com.enviro.assessment.grad001.kganyamaleka.DTO.WasteCategoryDTO;
 import com.enviro.assessment.grad001.kganyamaleka.entities.WasteCategory;
+import com.enviro.assessment.grad001.kganyamaleka.exceptions.ResourceNotFoundException;
 import com.enviro.assessment.grad001.kganyamaleka.repository.WasteCategoryRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -33,23 +34,19 @@ public class WasteCategoryService {
     /**
      * Retrieves a waste category by its ID.
      * @param id the ID of the waste category.
-     * @return an Optional containing the waste category, or empty if not found.
+     * @return the WasteCategoryDTO.
+     * @throws ResourceNotFoundException if the category is not found.
      */
     public WasteCategoryDTO getById(Long id) {
         WasteCategory category = repository.findById(id)
-                .orElse(null);  // Return null if not found
-        if (category != null) {
-            return new WasteCategoryDTO(category);  // Return DTO if found
-        } else {
-            return null;  // Return null if not found
-        }
+                .orElseThrow(() -> new ResourceNotFoundException("Waste category with ID " + id + " not found."));
+        return new WasteCategoryDTO(category);  // Return DTO
     }
-
 
     /**
      * Adds a new waste category.
      * @param category the waste category to be added.
-     * @return the saved waste category.
+     * @return the saved WasteCategoryDTO.
      */
     @Transactional
     public WasteCategoryDTO addCategory(WasteCategory category) {
@@ -61,16 +58,20 @@ public class WasteCategoryService {
      * Retrieves the total number of waste categories.
      * @return the count of all waste categories.
      */
-    public Long numberOfCategories(){
+    public Long numberOfCategories() {
         return repository.count();
     }
 
     /**
      * Deletes a waste category by its ID.
      * @param id the ID of the waste category to be deleted.
+     * @throws ResourceNotFoundException if the category is not found.
      */
     @Transactional
-    public void deleteCategoryByID(Long id){
+    public void deleteCategoryByID(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Cannot delete. Waste category with ID " + id + " not found.");
+        }
         repository.deleteById(id);
     }
 }
